+++ example/lib/pages/home_page.dart
@@ -4,7 +4,6 @@ import 'package:example/router/router.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:recase/recase.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class HomePage extends StatelessWidget {
@@ -64,38 +63,32 @@ class _StyleDropdownButton extends ConsumerWidget {
         child: DropdownButton<AdaptiveStyle>(
           value: ref.watch(adaptiveStyleProvider),
           items: AdaptiveStyle.values
+              // ignore: deprecated_member_use
+              .where((style) => style != AdaptiveStyle.cupertino)
               .map(
                 (style) => DropdownMenuItem(
                   value: style,
-                  child: Text(
-                    style.name.pascalCase,
-                  ),
+                  child: Text(style.label),
                 ),
               )
               .toList(),
           onChanged: (style) =>
-              ref.read(adaptiveStyleProvider.notifier).update((_) => style!),
+              ref.read(adaptiveStyleProvider.notifier).update(style!),
         ),
       ),
     );
   }
 }
 
-final adaptiveStyleProvider = StateProvider((ref) => AdaptiveStyle.adaptive);
+final adaptiveStyleProvider =
+    StateNotifierProvider<AdaptiveStyleNotifier, AdaptiveStyle>(
+  (ref) => AdaptiveStyleNotifier(),
+);
 
-extension AdaptiveStyleX on AdaptiveStyle {
-  TargetPlatform? get platform {
-    switch (this) {
-      case AdaptiveStyle.material:
-        return TargetPlatform.android;
-      // ignore: deprecated_member_use
-      case AdaptiveStyle.cupertino:
-      case AdaptiveStyle.iOS:
-        return TargetPlatform.iOS;
-      case AdaptiveStyle.macOS:
-        return TargetPlatform.macOS;
-      case AdaptiveStyle.adaptive:
-        return null;
-    }
+class AdaptiveStyleNotifier extends StateNotifier<AdaptiveStyle> {
+  AdaptiveStyleNotifier() : super(AdaptiveStyle.adaptive);
+  void update(AdaptiveStyle style) {
+    AdaptiveDialog.instance.updateConfiguration(defaultStyle: style);
+    state = style;
   }
 }