+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.6.0
 
 - Support for macOS dialogs
+- Add `AdaptiveDialog.instance.updateConfiguration`, a method used for changing default configuration
 
 ## 1.5.1